src/loadgenerator/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
backoff==1.10.0
    # via opentelemetry-exporter-otlp-proto-grpc
brotli==1.0.9
    # via geventhttpclient
certifi==2020.12.5
    # via
    #   geventhttpclient
    #   requests
charset-normalizer==2.0.10
    # via requests
click==7.1.2
    # via flask
configargparse==1.2.3
    # via locust
deprecated==1.2.13
    # via opentelemetry-api
faker==8.12.1
    # via -r requirements.in
flask==2.0.2
    # via
    #   flask-basicauth
    #   flask-cors
    #   locust
flask-basicauth==0.2.0
    # via locust
flask-cors==3.0.10
    # via locust
gevent==21.12.0
    # via
    #   geventhttpclient
    #   locust
geventhttpclient==1.5.3
    # via locust
googleapis-common-protos==1.54.0
    # via opentelemetry-exporter-otlp-proto-grpc
greenlet==1.1.2
    # via gevent
grpcio==1.43.0
    # via opentelemetry-exporter-otlp-proto-grpc
idna==2.10
    # via requests
itsdangerous==2.0.1
    # via flask
jinja2==3.0.3
    # via flask
locust==2.6.1
    # via -r requirements.in
markupsafe==2.0.1
    # via jinja2
msgpack==1.0.0
    # via locust
opentelemetry-api==1.8.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-urllib3
    #   opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc==1.8.0
    # via -r requirements.in
opentelemetry-instrumentation==0.27b0
    # via
    #   -r requirements.in
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-urllib3
opentelemetry-instrumentation-requests==0.27b0
    # via -r requirements.in
opentelemetry-instrumentation-urllib3==0.27b0
    # via -r requirements.in
opentelemetry-proto==1.8.0
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.8.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions==0.27b0
    # via
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-urllib3
    #   opentelemetry-sdk
opentelemetry-util-http==0.27b0
    # via
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-urllib3
protobuf==3.19.3
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
psutil==5.7.2
    # via locust
python-dateutil==2.8.2
    # via faker
pyzmq==22.3.0
    # via locust
requests==2.27.1
    # via
    #   -r requirements.in
    #   locust
roundrobin==0.0.2
    # via locust
six==1.15.0
    # via
    #   flask-cors
    #   geventhttpclient
    #   grpcio
    #   python-dateutil
text-unidecode==1.3
    # via faker
typing-extensions==4.0.1
    # via locust
urllib3==1.26.8
    # via
    #   -r requirements.in
    #   requests
werkzeug==2.0.2
    # via
    #   flask
    #   locust
wrapt==1.13.3
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-urllib3
zope.event==4.5.0
    # via gevent
zope.interface==5.1.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
